requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.0.0
scipy>=1.10.0
pathlib>=1.0.1
python-dotenv>=0.19.0
streamlit>=1.22.0
requests>=2.28.0
tqdm>=4.65.0
pyarrow>=12.0.0
fastparquet>=2023.1.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0 